+++ packages/schema/migrations/0003_spend_records.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+DROP TABLE IF EXISTS vs_spend_records;
+COMMIT;